Separate analyzer benchmark scenarios from core
Keep BenchmarkDotNet classes in AnalyzerBenchmarks.cs, move runner and compilation mechanics to AnalyzerBenchmarkInfrastructure.cs, and isolate analyzer registrations and source generation in AnalyzerScenarios.cs.

Update the authoring skill so ordinary analyzer additions touch only the scenario definitions.


Copilot-Session: 412a1a5f-d793-47bc-81f5-b341fd2b1cd7
